requirements.txt
azure-core==1.38.0
azure-iot-device==2.14.0
certifi==2026.1.4
charset-normalizer==3.4.4
deprecation==2.1.0
dotenv==0.9.9
idna==3.11
janus==2.0.0
packaging==25.0
paho-mqtt==1.6.1
pyserial==3.5
PySocks==1.7.1
python-dotenv==1.2.1
requests==2.32.5
requests-unixsocket2==1.0.1
typing_extensions==4.15.0
urllib3==2.6.3